refactor: move stack/sandbox ID logic into @aws-blocks/blocks/scripts
- Export getStackId() and getSandboxId() from @aws-blocks/blocks/scripts
- getStackId() reads stackId from .blocks/config.json (same file as telemetry)
- getSandboxId() generates/reads per-machine ID in .blocks-sandbox/sandbox-id
- Templates now just: import { getStackId, getSandboxId } from '@aws-blocks/blocks/scripts'
- Removed sandbox-id.ts from all templates (logic lives in core now)
- CLI writes stackId to .blocks/config.json during scaffolding
