+++ .changeset/derive-stack-name-from-package-json.md
@@ -1,7 +1,11 @@
 ---
 "@aws-blocks/create-blocks-app": patch
+"@aws-blocks/core": patch
 ---
 
-fix: generate unique stackId in blocks/config.json for CloudFormation stack naming
+fix: generate unique stackId in .blocks/config.json, export getStackId/getSandboxId from @aws-blocks/blocks/scripts
 
-Stack names are now derived from a `stackId` stored in `blocks/config.json`, generated at scaffold time as `<name>.slice(0,16)-<random6>`. This ensures unique stack names across apps in the same account/region and uses the same ID for both production (`<stackId>-prod`) and sandbox (`<stackId>-sandbox`) deployments. The `getSandboxId` script has been removed.
+Stack names are now derived from a `stackId` in `.blocks/config.json`, generated at scaffold time as `<name>.slice(0,16)-<random6>`. Templates import `getStackId()` and `getSandboxId()` from `@aws-blocks/blocks/scripts` — no more inline filesystem logic in `index.cdk.ts`.
+
+Production: `<stackId>-prod`
+Sandbox: `<stackId>-<username(8)>-<random(4)>` (per-machine, gitignored)
