+++ packages/core/src/scripts/index.ts
@@ -19,3 +19,4 @@ export {
   type BuildAndSendEventOptions,
 } from '../telemetry/index.js';
 export { telemetry, type TelemetryOptions } from './telemetry.js';
+export { getStackId, getSandboxId } from './stack-id.js';
