+++ packages/create-blocks-app/templates/auth-cognito/aws-blocks/index.cdk.ts
@@ -4,8 +4,7 @@ import { RemovalPolicies, Mixins } from 'aws-cdk-lib';
 import { Hosting, BlocksStack, SandboxDisableDeletionProtection } from '@aws-blocks/blocks/cdk';
 import { fileURLToPath } from 'node:url';
 import { dirname, join } from 'node:path';
-import { readFileSync } from 'node:fs';
-import { getSandboxId } from './scripts/sandbox-id.js';
+import { getStackId, getSandboxId } from '@aws-blocks/blocks/scripts';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -13,8 +12,7 @@ const app = new cdk.App();
 
 const sandboxMode = app.node.tryGetContext('sandboxMode') === 'true';
 
-const { stackId } = JSON.parse(readFileSync(join(__dirname, '..', 'blocks', 'config.json'), 'utf-8'));
-const stackName = sandboxMode ? `${stackId}-${getSandboxId()}` : `${stackId}-prod`;
+const stackName = sandboxMode ? `${getStackId()}-${getSandboxId()}` : `${getStackId()}-prod`;
 export const blocksStack = await BlocksStack.create(app, stackName, {
   backendHandlerPath: join(__dirname, 'index.handler.ts'),
   backendCDKPath: join(__dirname, 'index.ts')
